added base program

+++ src/marketplace/user/User.java
@@ -1,12 +1,12 @@
 package marketplace.user;
 
 public class User {
-  public long id;
+  public int id;
   public String firstName;
   public String lastName;
   public double amountOfMoney;
 
-  public User(long id, String firstName, String lastName, double amountOfMoney) {
+  public User(int id, String firstName, String lastName, double amountOfMoney) {
     this.id = id;
     this.firstName = firstName;
     this.lastName = lastName;
